requirements.txt
black
blinker
click
datetime
email-validator
flake8
Flask
flask-admin
flask_basicauth
flask_bootstrap
Flask-Login
flask-migrate
Flask-Mail
flask_mail
flask_security
Flask-SQLAlchemy
Flask-WTF
gunicorn
itsdangerous
Jinja2
MarkupSafe
pytest
pathlib
python-dotenv
SQLAlchemy==1.3.23
sqlalchemy_utils
Werkzeug
WTForms
requests
pyOpenSSL==19.0.0
pytz
psycopg2-binary
